Create Prims Algo.cpp
Prim's Algorithm is a greedy algorithm that is used to find the minimum spanning tree from a graph. This algorithm finds the subset of edges that includes every vertex of the graph such that the sum of the weights of the edges can be minimized.

It starts with the single node and explores all the adjacent nodes with all the connecting edges at every step. The edges with the minimal weights causing no cycles in the graph get selected.